examples/LLM/llama/ov_chat_app/docker/requirements.txt
transformers
llama-cpp-python
streamlit>=1.26.0
requests_futures
asyncio
aiohttp
accelerate
